SPOCAN-8 refactor menor, armo directorio ui para fragments y activity
