--- app/src/main/java/com/neiapp/spocan/ProfileActivity.java
@@ -1,4 +1,0 @@
-package com.neiapp.spocan;
-
-public class ProfileActivity {
-}
