+++ app/src/main/java/com/neiapp/spocan/ui/activity/InitiativeActivity.java
@@ -1,12 +1,13 @@
-package com.neiapp.spocan;
+package com.neiapp.spocan.ui.activity;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.neiapp.spocan.R;
 
 public class InitiativeActivity extends AppCompatActivity {
