+++ app/src/main/res/layout/fragment_configuration_.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".configuration_Fragment">
+    tools:context=".ui.fragment.ConfigurationFragment">
 
     <TextView
         android:id="@+id/textView6"
