+++ app/src/main/res/navigation/my_nav.xml
@@ -6,17 +6,17 @@
     app:startDestination="@id/profile_Fragment">
     <fragment
         android:id="@+id/profile_Fragment"
-        android:name="com.neiapp.spocan.Profile_Fragment"
+        android:name="com.neiapp.spocan.ui.fragment.ProfileFragment"
         android:label="fragment_profile_"
         tools:layout="@layout/fragment_profile_" />
     <fragment
         android:id="@+id/home_Fragment"
-        android:name="com.neiapp.spocan.Home_Fragment"
+        android:name="com.neiapp.spocan.ui.fragment.HomeFragment"
         android:label="fragment_home_"
         tools:layout="@layout/fragment_home_" />
     <fragment
         android:id="@+id/configuration_Fragment"
-        android:name="com.neiapp.spocan.configuration_Fragment"
+        android:name="com.neiapp.spocan.ui.fragment.ConfigurationFragment"
         android:label="fragment_configuration_"
         tools:layout="@layout/fragment_configuration_" />
 </navigation>
